+++ src/utils/network/Client.cpp
@@ -21,16 +21,20 @@ File Description:
 #include "utils/exception/ExceptionDefine.hpp"
 #include "utils/exception/basic/ErrorException.hpp"
 #include "utils/exception/basic/WarningException.hpp"
+#include "utils/exception/basic/NoneException.hpp"
 #include "utils/network/NetworkDefine.hpp"
 #include "utils/network/Client.hpp"
 #include "utils/verbose/Verbose.hpp"
 #include <sys/epoll.h>
 #include <unistd.h>
 #include <string.h>
+#include <iostream>
 #include <thread>
+#include <mutex>
 
 _cold void utils::network::Client::start(void)
 {
+    // Check status
     if (this->_status == utils::network::Status::Up) {
         throw utils::exception::WarningException(utils::exception::InternalCode::AlreadyRunning);
     } else if (this->_status == utils::network::Status::Terminated) {
@@ -64,7 +68,7 @@ _cold void utils::network::Client::start(void)
         }
 
         this->_status = utils::network::Status::Up;
-    } catch (...) {
+    } catch (const utils::exception::IException& e) {
         this->_status = utils::network::Status::Crashed;
         throw;
     }
@@ -72,7 +76,9 @@ _cold void utils::network::Client::start(void)
 
 _cold void utils::network::Client::stop(void)
 {
+    // Check status
     if (this->_status != utils::network::Status::Up) return;
+
     onBasicVerbose("Stopping client...");
     this->_socket->close();
     ::close(this->_epfd);
@@ -82,7 +88,9 @@ _cold void utils::network::Client::stop(void)
 
 _cold void utils::network::Client::kill(void)
 {
+    // Check status
     if (this->_status == utils::network::Status::Terminated) return;
+
     onBasicVerbose("Killing client...");
     this->_socket->close();
     if (this->_epfd != -1) ::close(this->_epfd);
@@ -106,7 +114,7 @@ _hot _nodiscard const utils::network::Payloads& utils::network::Client::listen(v
 
         if (res < 0) _unlikely {
             if (errno == EINTR) return this->_payloads; // handle the ctrl-c before the first connection
-            this->kill();
+            this->stop();
             this->_status = utils::network::Status::Crashed;
             throw utils::exception::ErrorException(utils::exception::InternalCode::Poll, strerror(errno));
         }
@@ -121,9 +129,14 @@ _hot _nodiscard const utils::network::Payloads& utils::network::Client::listen(v
         */
         if (events[0].events & (EPOLLERR | EPOLLHUP)) _unlikely {
             onBasicVerbose("Detected invalid epoll event, exiting...");
-            this->kill();
+            this->stop();
             this->_status = utils::network::Status::Crashed;
-            throw utils::exception::ErrorException(utils::exception::InternalCode::Poll);
+            if (events[0].events & EPOLLHUP) _likely {
+                onBasicVerbose("Socket closed...");
+                this->stop();
+                this->_status = utils::network::Status::Down;
+                return this->_payloads;
+            } else _unlikely {throw utils::exception::ErrorException(utils::exception::InternalCode::Poll);}
         }
 
         // Nothing to read
@@ -137,15 +150,14 @@ _hot _nodiscard const utils::network::Payloads& utils::network::Client::listen(v
         } catch (const utils::exception::IException& e) {
             if (e.isNone() && e.getCode() == utils::exception::InternalCode::SocketClosed) {
                 onBasicVerbose("Socket closed...");
+                this->stop();
                 this->_status = utils::network::Status::Down;
             } else {
                 onBasicVerbose("Error during the client request handling...");
-                onBasicVerbose(e.formated());
-                this->kill();
+                onDebugVerboseC(std::cerr, e.formated());
+                this->stop();
                 this->_status = utils::network::Status::Crashed;
             }
-            ::close(this->_epfd);
-            this->_epfd = -1;
             return this->_payloads;
         }
     }
@@ -170,7 +182,7 @@ _hot void utils::network::Client::join(void)
 
         if (res < 0) _unlikely {
             if (errno == EINTR) return; // handle the ctrl-c before the first connection
-            this->kill();
+            this->stop();
             this->_status = utils::network::Status::Crashed;
             throw utils::exception::ErrorException(utils::exception::InternalCode::Poll, strerror(errno));
         }
@@ -185,9 +197,14 @@ _hot void utils::network::Client::join(void)
         */
         if (events[0].events & (EPOLLERR | EPOLLHUP)) _unlikely {
             onBasicVerbose("Detected invalid epoll event, exiting...");
-            this->kill();
+            this->stop();
             this->_status = utils::network::Status::Crashed;
-            throw utils::exception::ErrorException(utils::exception::InternalCode::Poll);
+            if (events[0].events & EPOLLHUP) _likely {
+                onBasicVerbose("Socket closed...");
+                this->stop();
+                this->_status = utils::network::Status::Down;
+                return;
+            } else _unlikely {throw utils::exception::ErrorException(utils::exception::InternalCode::Poll);}
         }
 
         // Nothing to read
@@ -204,9 +221,17 @@ _hot void utils::network::Client::flush(void)
     if (this->_status != utils::network::Status::Up) _unlikely {return;}
 
     try {this->_socket->flush();}
-    catch (...) {
-        this->kill();
-        this->_status = utils::network::Status::Crashed;
+    catch (const utils::exception::IException& e) {
+        if (e.isNone() && e.getCode() == utils::exception::InternalCode::SocketClosed) _likely {
+            onBasicVerbose("Socket closed...");
+            this->stop();
+            this->_status = utils::network::Status::Down;
+        } else _unlikely {
+            onBasicVerbose("Error detected...");
+            onDebugVerboseC(std::cerr, e.formated());
+            this->stop();
+            this->_status = utils::network::Status::Crashed;
+        }
         throw;
     }
 }
@@ -218,9 +243,17 @@ _hot void utils::network::Client::send<false>(const utils::network::Payload& pay
     if (this->_status != utils::network::Status::Up) _unlikely {return;}
 
     try {this->_socket->send(payload);}
-    catch (...) {
-        this->kill();
-        this->_status = utils::network::Status::Crashed;
+    catch (const utils::exception::IException& e) {
+        if (e.isNone() && e.getCode() == utils::exception::InternalCode::SocketClosed) _likely {
+            onBasicVerbose("Socket closed...");
+            this->stop();
+            this->_status = utils::network::Status::Down;
+        } else _unlikely {
+            onBasicVerbose("Error detected...");
+            onDebugVerboseC(std::cerr, e.formated());
+            this->stop();
+            this->_status = utils::network::Status::Crashed;
+        }
         throw;
     }
 }
@@ -232,9 +265,17 @@ _hot void utils::network::Client::send<true>(const utils::network::Payload& payl
     if (this->_status != utils::network::Status::Up) _unlikely {return;}
 
     try {this->_socket->sendBuffered(payload);}
-    catch (...) {
-        this->kill();
-        this->_status = utils::network::Status::Crashed;
+    catch (const utils::exception::IException& e) {
+        if (e.isNone() && e.getCode() == utils::exception::InternalCode::SocketClosed) _likely {
+            onBasicVerbose("Socket closed...");
+            this->stop();
+            this->_status = utils::network::Status::Down;
+        } else _unlikely {
+            onBasicVerbose("Error detected...");
+            onDebugVerboseC(std::cerr, e.formated());
+            this->stop();
+            this->_status = utils::network::Status::Crashed;
+        }
         throw;
     }
 }