+++ src/utils/verbose/Verbose.cpp
@@ -8,7 +8,7 @@
  ╚═╝  ╚═╝╚═╝  ╚═╝╚═╝  ╚═╝   ╚═╝   ╚═╝  ╚═╝╚═╝  ╚═══╝╚═╝╚═╝  ╚═╝
 
 Edition:
-##  @date 16/07/2026 by @author Tsukini
+##  @date 16/08/2026 by @author Tsukini
 
 File Name:
 ##  @file Verbose.cpp
@@ -20,8 +20,8 @@ File Description:
 #include "utils/verbose/Verbose.hpp"
 #include <mutex>
 
-// Declaration of the global var for the security on stdout writing
-std::mutex utils::verbose::stdout_lock;
+// Declaration of the global var for the security on output writing
+std::mutex utils::verbose::output_lock;
 
 // Declaration of the global var for the verbose mode
 volatile utils::verbose::Verbose utils::verbose::verbose = utils::verbose::Verbose::Basic;